# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handler.ChangePasswordInput:
    properties:
      current_password:
        type: string
      new_password:
        minLength: 8
        type: string
    required:
    - current_password
    - new_password
    type: object
...
  handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handler.LinkGitHubInput:
    properties:
      github_code:
        type: string
    required:
    - github_code
    type: object
  handler.MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  handler.UnlinkGitHubInput:
    properties:
      password:
        type: string
    required:
    - password
    type: object
  handler.UpdateUserProfileInput:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      name:
        type: string
    type: object
  handler.UserProfileResponse:
    properties:
      avatar_url:
        type: string
      email:
        type: string
      github_id:
        type: integer
      id:
        type: string
      name:
        type: string
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: This is a sample server celler server.
  title: Swagger Memoria API
  version: "1.0"
paths:
...
  /users/account:
    delete:
      consumes:
      - application/json
      description: Полностью удаляет аккаунт текущего пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Аккаунт успешно удален
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Удалить аккаунт
      tags:
      - users
  /users/github/link:
    post:
      consumes:
      - application/json
      description: Привязывает GitHub аккаунт к текущему пользователю
      parameters:
      - description: Код авторизации GitHub
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.LinkGitHubInput'
      produces:
      - application/json
      responses:
        "200":
          description: GitHub аккаунт успешно привязан
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: GitHub аккаунт уже привязан к другому пользователю
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Привязать GitHub аккаунт
      tags:
      - users
  /users/github/unlink:
    post:
      consumes:
      - application/json
      description: Отвязывает GitHub аккаунт от текущего пользователя
      parameters:
      - description: Пароль для подтверждения
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.UnlinkGitHubInput'
      produces:
      - application/json
      responses:
        "200":
          description: GitHub аккаунт успешно отвязан
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Отвязать GitHub аккаунт
      tags:
      - users
  /users/password:
    put:
      consumes:
      - application/json
      description: Изменяет пароль текущего пользователя
      parameters:
      - description: Данные для изменения пароля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.ChangePasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: Пароль успешно изменен
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Изменить пароль
      tags:
      - users
  /users/profile:
    get:
      description: Возвращает профиль текущего аутентифицированного пользователя
      produces:
      - application/json
      responses:
        "200":
          description: Профиль пользователя
          schema:
            $ref: '#/definitions/handler.UserProfileResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Получить профиль пользователя
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Обновляет данные профиля текущего пользователя
      parameters:
      - description: Данные для обновления профиля
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateUserProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: Профиль успешно обновлен
          schema:
            $ref: '#/definitions/handler.MessageResponse'
        "400":
          description: Неверные входные данные
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Пользователь не найден
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Email уже используется
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "500":
          description: Ошибка сервера
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Обновить профиль пользователя
      tags:
      - users
securityDefinitions:
  apiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"